Measure and layout properly
Fulfil the onMeasure contract by allowing AT_MOST and UNSPECIFIED specs
instead of failing at runtime. In addition, use the view's padding to
make child views smaller, the way it should be. There is one problem,
which is that the ViewFlowIndicator implementations assume that the
scrolled region is the whole width of the container, whereas it might
actually be smaller, so they show wrong positions if left/right padding
is non-zero. This will probably require an interface change to fix.