Avoid views flickering when interrupting a scroll
If you start a new scroll while the scroll animation is in progress,
onInterceptTouchEvent will call mScroller.abortAnimation,
and when computeScroll is next called, postViewSwitched will run.

As computeScroll is called within View.draw, this is not a good time
to be fiddling with the layout by adding and removing children. It
causes the wrong view to be drawn for a frame when this happens.

Instead of calling postViewSwitched immediately,
I've changed it to be posted to the handler so it will be called
after this frame has finished drawing and before the next is laid out.